Reorganized 'schedulers/' module; Move CPDEL logic into HRevolve class, where it belongs
